// cart/sweatshop/components/scene3d/types.ts
export type Vec3 = [number, number, number];
// ...
export interface PointLightProps        { position?: Vec3; color?: string; intensity?: number; range?: number }